requirements.txt
flask
python-dotenv
gunicorn
requests
flask-cors
flask-sqlalchemy   